% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleJasparPredict.R
\name{singleJasparPredict}
\alias{singleJasparPredict}
\title{Title}
\usage{
singleJasparPredict(
  rsID,
  chr,
  position,
  ALT,
  REF,
  window,
  strand,
  threshold,
  plotLogo
)
}
\arguments{
\item{rsID}{SNP的rs号}

\item{chr}{SNP所在染色体号}

\item{position}{SNP所在的碱基位置}

\item{ALT}{风险等位基因}

\item{REF}{参考等位基因}

\item{window}{SNP上下游扩展序列的窗口大小}

\item{strand}{基因所在链的正反}

\item{threshold}{JASPAR预测显著结合的阈值}

\item{plotLogo}{是否绘制转录因子Motif的LOGO图}
}
\description{
Title
}
